+++ lib/lightgbm.rb
@@ -10,6 +10,7 @@ require "lightgbm/version"
 # scikit-learn API
 require "lightgbm/model"
 require "lightgbm/classifier"
+require "lightgbm/ranker"
 require "lightgbm/regressor"
 
 module LightGBM