+++ lib/lightgbm/classifier.rb
@@ -4,7 +4,7 @@ module LightGBM
       super
     end
 
-    def fit(x, y, categorical_feature: "auto", early_stopping_rounds: nil, verbose: true)
+    def fit(x, y, eval_set: nil, categorical_feature: "auto", early_stopping_rounds: nil, verbose: true)
       n_classes = y.uniq.size
 
       params = @params.dup
@@ -16,10 +16,13 @@ module LightGBM
       end
 
       train_set = Dataset.new(x, label: y, categorical_feature: categorical_feature)
+      valid_sets = Array(eval_set).map { |v| Dataset.new(v[0], label: v[1], reference: train_set) }
+
       @booster = LightGBM.train(params, train_set,
         num_boost_round: @n_estimators,
         early_stopping_rounds: early_stopping_rounds,
-        verbose_eval: verbose
+        verbose_eval: verbose,
+        valid_sets: valid_sets
       )
       nil
     end