+++ lib/lightgbm.rb
@@ -8,7 +8,9 @@ require "lightgbm/dataset"
 require "lightgbm/version"
 
 # scikit-learn API
+require "lightgbm/model"
 require "lightgbm/classifier"
+require "lightgbm/ranker"
 require "lightgbm/regressor"
 
 module LightGBM