+++ lib/lightgbm/dataset.rb
@@ -2,7 +2,7 @@ module LightGBM
   class Dataset
     attr_reader :data, :params
 
-    def initialize(data, label: nil, weight: nil, params: nil, reference: nil, used_indices: nil, categorical_feature: "auto")
+    def initialize(data, label: nil, weight: nil, group: nil, params: nil, reference: nil, used_indices: nil, categorical_feature: "auto")
       @data = data
 
       # TODO stringify params
@@ -42,8 +42,9 @@ module LightGBM
       end
       ObjectSpace.define_finalizer(self, self.class.finalize(handle_pointer)) unless used_indices
 
-      set_field("label", label) if label
-      set_field("weight", weight) if weight
+      self.label = label if label
+      self.weight = weight if weight
+      self.group = group if group
     end
 
     def label
@@ -52,6 +53,18 @@ module LightGBM
 
     def weight
       field("weight")
+    end
+
+    def label=(label)
+      set_field("label", label)
+    end
+
+    def weight=(weight)
+      set_field("weight", weight)
+    end
+
+    def group=(group)
+      set_field("group", group, type: :int32)
     end
 
     def num_data
@@ -105,11 +118,17 @@ module LightGBM
       out_ptr.read_pointer.read_array_of_float(num_data)
     end
 
-    def set_field(field_name, data)
+    def set_field(field_name, data, type: :float)
       data = data.to_a unless data.is_a?(Array)
-      c_data = ::FFI::MemoryPointer.new(:float, data.count)
-      c_data.put_array_of_float(0, data)
-      check_result FFI.LGBM_DatasetSetField(handle_pointer, field_name, c_data, data.count, 0)
+      if type == :int32
+        c_data = ::FFI::MemoryPointer.new(:int32, data.count)
+        c_data.put_array_of_int32(0, data)
+        check_result FFI.LGBM_DatasetSetField(handle_pointer, field_name, c_data, data.count, 2)
+      else
+        c_data = ::FFI::MemoryPointer.new(:float, data.count)
+        c_data.put_array_of_float(0, data)
+        check_result FFI.LGBM_DatasetSetField(handle_pointer, field_name, c_data, data.count, 0)
+      end
     end
 
     def matrix?(data)