+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.4 [unreleased]
 
 - Friendlier message when LightGBM not found
+- Added `Ranker`
 - Free memory when objects are destroyed
 - Removed unreleased `dump_text` method