+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.1.4 [unreleased]
 
 - Friendlier message when LightGBM not found
+- Added `Ranker`
+- Added early stopping to Scikit-Learn API
 - Free memory when objects are destroyed
 - Removed unreleased `dump_text` method